function2 update

+++ calculation2.java
@@ -1,5 +1,5 @@
 public class calculation2 {
-	public int function2(int n){
+	public double function2(int n){
 		double res=1;
 		int i = 1;
 		while(i<=n){
